Applies review suggestions - changes camera related tests subclass to its own class
