+++ package/com.unity.formats.usd/Tests/Common/TestUtilityFunction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c792624bf7fca1409249d449c552913
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
