+++ package/com.unity.formats.usd/Tests/Editor/ExportCameraTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b24d3c8d0e5f2f84e95175f0cf0a1114
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
